Fix some typing and stuff
