+++ apps/web/src/components/forms/multi-step/steps/step-2-business-model/__tests__/schema.test.ts
@@ -1,9 +1,9 @@
 import { describe, expect, it } from 'vitest'
-import { 
-  type BusinessModelData, 
-  businessModelSchema, 
-  type SubsidiaryData, 
-  subsidiarySchema 
+import {
+  type BusinessModelData,
+  businessModelSchema,
+  type SubsidiaryData,
+  subsidiarySchema,
 } from '../schema'
 
 describe('Subsidiary Schema', () => {
