+++ apps/web/src/components/forms/multi-step/steps/step-4-sustainability-practices/index.ts
@@ -1,2 +1,2 @@
 // Export sustainability practices components and schema
-export { sustainabilitySchema, sustainabilityFields, sustainabilityStepConfig } from './schema'
+export { sustainabilityFields, sustainabilitySchema, sustainabilityStepConfig } from './schema'
